requirements.txt
aiogram==2.21
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
asynccpu==1.2.2
asyncffmpeg==1.2.0
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
decorator==4.4.2
exif==1.3.5
ffmpeg-python==0.2.0
frozenlist==1.3.0
future==0.18.2
idna==3.3
imageio==2.19.5
imageio-ffmpeg==0.4.7
moviepy==1.0.3
multidict==6.0.2
numpy==1.23.1
opencv-python==4.6.0.66
piexif==1.1.3
Pillow==9.2.0
plum-py==0.8.1
proglog==0.1.10
psutil==5.9.1
pytz==2022.1
pywin32==304
PyYAML==6.0
requests==2.28.1
tqdm==4.64.0
urllib3==1.26.10
yarl==1.7.2
